chore(api): consolidate files and remove redundant types file
