+++ packages/blockchain-wallet-v4-frontend/src/modals/Brokerage/Banks/Withdraw/EnterAmount/index.tsx
@@ -3,7 +3,7 @@ import { connect, ConnectedProps } from 'react-redux'
 import { bindActionCreators, Dispatch } from 'redux'
 
 import { Remote } from 'blockchain-wallet-v4/src'
-import { SBPaymentTypes } from 'blockchain-wallet-v4/src/network/api/settingsComponent/types'
+import { SBPaymentTypes } from 'blockchain-wallet-v4/src/network/api/simpleBuy/types'
 import {
   BeneficiaryType,
   ExtractSuccess,
