+++ packages/blockchain-wallet-v4/src/network/api/custodial/index.ts
@@ -1,5 +1,5 @@
 import { CoinType, SBPaymentTypes, WalletFiatType } from 'core/types'
-import { BankTransferAccountType } from 'data/types'
+import { BankTransferAccountType, ProductEligibility } from 'data/types'
 
 import {
   BeneficiariesType,
@@ -101,9 +101,16 @@ export default ({ authorizedGet, authorizedPost, nabuUrl }) => {
       data: request
     })
 
+  const getProductsEligibility = (): ProductEligibility[] =>
+    authorizedGet({
+      url: nabuUrl,
+      endPoint: '/eligible/products'
+    })
+
   return {
     checkWithdrawalLocks,
     getBeneficiaries,
+    getProductsEligibility,
     getWithdrawalLocks,
     getWithdrawalFees,
     initiateCustodialTransfer,
