+++ packages/blockchain-wallet-v4/src/network/api/index.ts
@@ -16,7 +16,6 @@ import misc from './misc'
 import profile from './profile'
 import rates from './rates'
 import settings from './settings'
-import settingsComponent from './settingsComponent'
 import simpleBuy from './simpleBuy'
 import swap from './swap'
 import wallet from './wallet'
@@ -81,11 +80,6 @@ const api = ({
       ...http
     }),
     ...settings({ rootUrl, ...http }),
-    ...settingsComponent({
-      nabuUrl,
-      authorizedGet: authorizedHttp.get,
-      ...http
-    }),
     ...simpleBuy({
       everypayUrl,
       nabuUrl,
