--- packages/blockchain-wallet-v4/src/network/api/settingsComponent/index.ts
@@ -1,13 +1,0 @@
-import { ProductEligibility } from 'data/types'
-
-export default ({ authorizedGet, nabuUrl }) => {
-  const getProductsEligiblity = (): ProductEligibility[] =>
-    authorizedGet({
-      url: nabuUrl,
-      endPoint: '/eligible/products'
-    })
-
-  return {
-    getProductsEligiblity
-  }
-}
